TST: Add eq_repr comparison helper
The escape codes in test error messages get processed by the terminal,
which can result in a failing test reporting that green "foo" isn't
equal to green "foo" because of a hidden difference in the escape
codes.  Or worse, in the case of a clear screen code, a failing test
clears the screen.

This helper prevents that.