ENH+RF: Drop id_column argument, support multi-column id
Make Tabular.rewrite take a dictionary with {column: value, ...} to
identify the row to rewrite.  This way, we don't force the caller
specify the id column at instance creation.  Also, we can easily
support a multi-column id without having to worry about making
__init__'s id_column and rewrite's ids take both single and
multi-value forms.

The tradeoff is that the rewrite call is a little more verbose for
simple cases, but the intention should also be clearer (the
column->value mapping is right there so you don't have to think "oh
yeah, that first value is a value for id column").

Re: #14