ENH: Support extracting column names from the data
Get the names from the keys attribute.  This is intended to work with
OrderedDict's, but we might as well support any object with keys
rather than raising an error.  When callers use regular dicts, the
output will make it obvious that the columns aren't sorted.

Make __init__'s columns parameter optional rather than removing it
entirely because 1) we don't want to force callers to use an
OrderedDict, 2) the columns argument enables the user to select a
subset of the keys, and 3) __call__ may be extended to support
sequences, in which case the columns parameter would be necessary.

Closes #2.