ENH: Add an option to include a header in the output
The two main decisions here are 1) the interface for triggering the
inclusion of the header and 2) how the header for a column should
share styling with the values for the column.

For 1, a header is added if the style contains a "header_"
key (trailing underscore to avoid conflicts with column names).

For 2, the header style inherits the columns alignment and width but
not other attributes (color, bold, underline).  The value of the
style's header_ key can be used to provide header specific
styling (e.g., all make all headers bold).

Closes #3.