ENH: Support value->{color,bold,underline} mapping
In addition to simple column styles like {"bold": True} and {"color":
"green"}, allow the style attribute to be determined based on the
field value.  We have to handle color differently than bold and
underline because color's value is the attribute of interest, whereas
the value for bold and underline is just a boolean switch.

Re: #1