TermProcessors: Change method for deciding to add a reset code
To decide whether to add a terminal reset code, we test whether the
result minus any flanking white space is equal to the value being
rendered.  At the moment, the only processors that change the string
content are the ones that should end with a reset code, so this works
for now, but it isn't very robust.  If we, for example, add a
processor for truncating field values, it will break.

Instead add a reset code of the result contains an escape character.