RF: Share fields instance among different styles
We recreate field instances for overriding styles (including headers),
but this makes it difficult to share information with these fields,
which we need to do if we want to resize widths.  Instead of creating
separate instances, let's share the same Field instances across all
fields and add different lists of processors.  Processors in the
"core" list are always called; processors in the "default" list are
called by default, but this can be overridden with a different set of
processors.

A side-effect of this change is that "width" and "alignment" are
shared among all fields and can no longer be overridden through
Tabular.__call__'s style parameter.