RF: Change approach for repaint
Instead of clearing the screen, move up a number of lines equal to the
length of rows and then rewrite the rows.  It is OK if some number of
the top rows have scrolled off the screen because in that case the
move-up codes will be noops, and the rewrite of rows will "restore"
the position.

Since this approach no longer clears the screen, this change makes it
possible for a future commit to update __call__ to use the repaint
method.