RF: Use dict for label and interval values
These used to be tuples where the first item was either "interval" or
"label" (although of course list could also be used).  With the recent
jsonschema validation changes, they are required to be lists.

I preferred the tuples to dicts because there is only one supported
key.  But I prefer a single-key dictionary to a heterogeneous list,
particularly for the interval type.  The interval type still has
another heterogeneous list in its definition, but changing to a dict
in this case seems a bit unwieldy because you'd have to define
something like {"lower": M, "upper": N, "value": "v"} for each item.