RF: StyleProcessors: Change 'translate' method to 'render'
The translate method assumes that we can take a style key like bold,
translate it into a specific output code (e.g., the terminal ansi code
for bold), and then always just prepend this to the actual field
value.

This `translate(key) + value` approach, though, leaks details about
terminal-specific rendering into the StyleProcessors method.  Instead,
the method should also consider the value to be rendered, so that it
could, for example, wrap it with an appropriate prefix and suffix.