DOC: elements: Mark align and width as "column" scope
There are currently two defined scopes, "table" and "field".  "table"
is for things like the separator, which conceptually is an attribute
of the entire table.  "field" is for things like the color that can
change for any field (i.e., a particular [row, col] location).

"align" and "width" don't fit into either of these because we force a
column to share the same width and alignment, so label them as
"column" scope.

Currently, these scope labels are just for documentation, but we might
want to check them when Tabular is passed an overriding style.