RF: field: Split processors into pre and post formatting
We haven't needed to distinguish between these yet, but, if we allow
the style to define arbitrary functions, we will want to pass the
function the value before it's formatted or otherwise processed.  This
way, the user receives the data in the same form as what they passed
to Tabular.