RF: tabular: Add concept of processor "group"
Before we write a row, _set_widths raises a RewritePrevious if any of
the previous rows looks in need of an update.  To prepare this check
for arbitrary transform function via style, we need to update it to
check the width of that function's result not the raw field value.
For example, if we use humanize.naturalsize, we'd want to set the
width for "3 Bytes" not for "3".  On top of that, we need to make sure
the transform function isn't called on the header or when there is an
overriding style.

Deal with this by introducing "default" and "override" proc_groups and
teach _set_widths to choose the preformat-processor keys based on that
value.  This makes writing a row a little more verbose, so add a
helper function, _check_and_write, that takes care of doing the "get
processor group -> set widths -> maybe repaint -> write row" routine.