RF: tabular: Rename "transform method" to "normalizer"
Tabular's _transform_method takes any of the acceptable forms for a
data row (dict, sequence, other object + attrs) and transforms it to a
dict for internal use.  By doing this transformation soon after we
receive the data, most of the internal methods only have to worry
about handling a dict.

After the _transform_method was added, we introduced an unrelated
"transform" style element.  To avoid any confusion, refer to the
former as a "normalizer".