BF: Raise StyleFunctionError with original traceback
We can't rely on the transform function's error being shown in the
traceback "above" because this depends on the caller's handling of the
StyleFunctionError.  To make StyleFunctionError easier to debug under
various handling scenarios, 1) reraise with the original traceback and
2) add the original exception type and message to StyleFunctionError's
message.  In case the original traceback doesn't come through, 2
should at least give some hint at what's going wrong.

The test demonstrates one case that, with Python 2, would fail before
this patch (i.e., the original traceback information was lost).

Closes #43.