RF: TermProcessors: Remove _maybe_reset method
The idea here was to only add a single trailing reset code, not one
for each ansi code at the start of the string.  However, we're going
to modify "underline" so that it doesn't underline flanking
whitespace.  For that, we'll need an internal reset, which would mean
every time we underline we'd have two resets.  So, given we'll have
two resets one way or the other, remove the complication of
_maybe_reset.