RF: Move content handling out of Tabular
Introduce a new class that is responsible for handling the
row->content conversion.  This means that Tabular no longer needs to
worry about managing StyleFields or rows.  After setting up a Content
instance, its only interaction with it is through the "update" method.

While this simplifies Tabular in its current state, the main
motivation for adding the Content layer is to make adding summary rows
cleaner.  Now Tabular shouldn't have to care about whether there are
summary rows.  It just knows that it has content of a given length and
that it needs to either append or overwrite (either a single line or
everything).

As a part of this restructuring, there is no longer an explicit
rewrite method.  Rewriting is now done when a set of column ids has
already been encountered.

Closes #20.