BF: field: Overwrite the processor lists
The idea was that you could use multiple `add` calls to build up the
list.  However, because StyleFields is using the "override" key for
anything "non-default", this could result in the style for one call
bleeding into the style for a future, unrelated one.  (I'm being lazy
and not adding a regression test for this because it will be covered
once summary lines are added.)  We could add an optional argument that
controls whether the list is overridden or extended, but since at the
moment we only ever use a single call for a given key and kind, let's
just always override the previous list.