RF: StyleFields: Add a compose method
The header_ style is a weird case because it needs to take some style
elements from each column (like width), but it should also be possible
to share an attribute across the header for each column (e.g., bold).

Move the logic for building this composite style into a method because
we're going to need to do the same thing for the style of summary
lines.