BF: tabular: Remove stale lines when content shrinks
Now that we allow summary functions, which can return multiple lines,
it's possible that the number of lines shrinks.  For example, say a
summary function returns a list value counts (one item per unique
value), resulting in a number of lines equal to the number of unique
values.  A callback function could then update a value to remove the
last occurrence of the value, resulting in one fewer line.