TST: tabular: Reduce boilerplate code
For all the Tabular tests, we patch pyout.tabular.Terminal and set the
stream to StringIO instance.  We also usually force_styling, but
sometimes that's dropped if it isn't relevant for the particular test.

Instead of doing these things in so many places, create a subclass of
pyout.Tabular that handles them.