BF: Truncate fields to avoid exceeding terminal width
If the width of a row exceeds the terminal's width, it wraps into the
next line and breaks all of our line counting-based modifications.  To
deal with this, we need truncate fields so that the row is within the
available width.  Note that we can't simply truncate at the row-level
because each field by include escape sequences.

This raises the issue of how we assign the free width to different
fields.  Use the following strategy: process fields with smaller
widths first so that they have a chance to claim free space before the
large ones claim more.  This should decrease the chances of a field
being entirely hidden.  We might want to revisit this assignment
strategy, but it's probably good enough.  If users are dealing with
columns that contain very long values, they can set the max widths of
those columns.