RF: StyleFields: Always truncate fields
The previous commit was the minimal code change required to switch
over to the Truncater processor class.  But now that we can modify the
truncation length easily, we can unconditionally truncate, keeping the
truncation length in sync with either the field length or the
explicitly set maximum width (whichever is smaller).

The next commit will build on this to truncate based on the total
table width available.