RF: tabular: Move terminal-independent bits to a base class
Following up on the last commit, remove the remaining
terminal-independent bits of Tabular.  This change highlights that
Tabular isn't a great name.  "Tabular" is probably a more suitable
name for the new Writer class since it expects to be writing tabular
data, while the current Tabular is terminal-specific and should
probably have "terminal" somewhere in its name.  At the same time, the
terminal is likely to always be the primary output, even if we do add
other support for other ones, so perhaps it makes sense use "Tabular"
for the main user-facing interface.  Dunno.

Note that this commit is nearly all pure code movement.