BF: interface: Prevent callbacks trying to update offscreen
If a line that is to be updated by a callback scrolls offscreen as the
output grows, we try to move back to that line with the appropriate
number of cuu1's, overwrite the line, and move back to the end of the
table with the appropriate number of cud1.  But, when called at the
top of the visible screen, cuu1 is undefined and likely a noop [*], so
we overwrite the wrong line and then move back with a number of cud1's
that overshoots the bottom of the table, producing blank lines.

Ideally we would be able to move back and update the offscreen line,
but it's not clear if this is easily doable.  At least until a better
solution is found, calculate whether a line is offscreen and repaint
the entire table if it is.

[*] https://www.gnu.org/software/termutils/manual/termcap-1.3/html_node/termcap_26.html#SEC26

Fixes #42.