RF: interface: Define a height property
As a first step to preventing attempts to update rows outside the
visible screen, we need to know how much visible space we have
available.

It'd probably be a good idea to keep checking term.height to catch
changes in the terminal size, but we don't do that for width yet, and
it's not clear how expensive doing so would be, so punt on that for
now.