RF: truncate: Tweak _truncate_right helper
This is mostly just cosmetics, but there's one functional difference.
We no longer make sure the marker is replacing more than just blank
characters.  This isn't actually necessary because Field will only pad
up to `length`, so the guard at the top of the function means that
there are usually more than blank characters.  The one case where
there could be only blank characters is if a user-specified transform
function adds them, in which case we *should* indicate that we're
eliding them.