field: Rename and clarify render's 'key' parameter
Describing this parameter as a style "key" is confusing because there
are two types of keys: those with values that are boolean
switches ("bold", "underline") and those with a set of
values (currently only "colors").  For "color", the style value (e.g.,
"blue") actually gets passed rather than the style key.