field: Unify signatures of processor-creating functions
The processor-creating by_key(), by_lookup(), and by_interval_lookup()
all have a 'key' parameter that is "underline" or "bold" (which have
boolean values) or a specific color for the "color" style key.  Aside
from the confusingly loose use of "key" here, post_from_style() is
responsible for knowing how to prepare the arguments for each
function, leading to logic that is spread out and hard to follow.

Adjust all the processor-creating functions to have the same signature
where their 'key' parameter is actually the style key ("underline",
"bold", "color"), and the processor-creating function is responsible
for checking the style key type.  This keeps the type logic closer to
the value lookup and simplifies post_from_style().