field: Support styling values by regular expression match
Do so by adding a new style attribute.  Another option would be to
extend "lookup" with some sort of "re" switch.  However, doing an
_ordered_ matching of regular expressions doesn't fit with the
dictionary value of "lookup".  Or we could drop "lookup" in favor of
"re_lookup", because anything that could be done with "lookup" can now
be done with "re_lookup".  But let's keep it because the lookup
performance is worse, the list syntax is more cumbersome and less
natural than the dictionary value, and it's nice to have a lookup
variant where the user doesn't have to worry about escaping regular
expressions.

Re: #46