re_lookup: Allow user to specify re.search flags
The main motivation for this is to allow case-insensitive matches, but
we might as well support all possible flags.

The advantage of adding a separate style attribute---as opposed to
somehow extending re_lookup's value to allow flags to be
specified---is that the user can then specify flags that apply to all
re_lookup values in the style by defining re_flags under default_.
With this approach, it's not possible to specify flags for a
particular regexp within the list, so we may want to rethink this
choice if the need for per-regexp flags arises.