Define a "no update, no styling" replacement for Tabular
We don't support Windows.  In order to do so, we'd need to (1) define
an interface that doesn't use blessings, (2) add support for rewriting
previous lines, and (3) add support for styling output.  We could at
least get some of the way by directly writing ANSI codes and using
colorama.  But that is for another time, and probably for someone else
who cares about supporting Windows.

Instead, do no. 1 by adding a mostly useless replacement for Tabular
that doesn't support updating or styling, with the unambitious goal of
having something that doesn't crash on Windows.

Re: #4