re_lookup: Guard against non-string values
The processor doesn't necessarily receive a string value.  Some common
cases are Nothing() objects for asynchronous values and numbers.
Pass the result for non-string values back untouched.

Another option would be to convert it to a string.  That could lead to
unintentional matches on str(object), but at the same time, that _is_
what will be shown, so perhaps it wouldn't be surprising.  For now, go
with the stricter route, adding a test that demonstrates the behavior.
We can revisit it if needed.