interface: Provide callers access to row values
Callers usually know the row values because they are passing them in
directly.  However, if values are being populated asynchronously (say
a function that connects to a remote and checks the status), callers
don't have an easy way to access them even though the values may be of
interest.

Give callers access to a row's value via __getitem__().