interface: A secondary initialization for Writer children
Instead of relying on the children setting up attributes, require them
to call a method after Writer has initialized.  The main motivation
for this restructuring is to make it easier for the Tabular class to
access the values of the upcoming "stream" and "interactive"
parameters.  To set up the stream interface, Tabular will want to
inspect these parameters _after_ Writer.__init__ has processed them.

This organization and interaction between classes still has a smell to
it, but these changes shouldn't make things worse.