field: Simplify PlainProcessors approach
Instead of rendering the values as is, set style_types to an empty
dictionary to avoid calling by_key() and friends at all.  This also
arguably makes it clearer how PlainProcessors is "plain" because it is
probably harder for the reader to connect render() to the "bold",
"color", and "underline" types.