Support hiding columns
pyout requires that all columns are specified before writing begins.
If none are explicitly specified, it infers the list of columns from
the first record received.

For some uses cases, it'd be desirable to expand the list of columns
as new keys are encountered (gh-49).  Unfortunately many parts of
pyout assume that the set of columns is fixed, so adding support for
auto-expanding the column list seems like it will require pretty deep
changes.  The issue is complicated by the fact that the request
involves more than being able to give records with new keys when
calling Tabular: asynchronous callbacks, themselves specified as
values to some column(s), should also be able to create new columns.

One of the motivations for auto-expanding the column list is the
desire to not show columns unless they actually contain information.
Address this by allowing columns to be hidden until they contain
something other than a Nothing() value.  Because it's easy to do and
could end up being useful (at least for debugging), also allow columns
to be unconditionally hidden.

The other motivation for auto-expanding columns is to let the caller
avoid specifying (or perhaps even knowing) the column names up front.
Hiding of course doesn't help with this.

Note that are a couple of details that make this a little more
complicated than simply skipping the hidden fields when rendering:

  * The header and summary should not be able to trigger unhiding.

  * After unhiding a column we need to reset fixed values that were
    calculated in build()

Re: #49, #84