common: Don't render zero-width fields
If a field has a width of zero because all of its values are
zero-width and its header isn't displayed (or is also zero-width), we
still render the field, flanking the column with separators.  That's
not a big deal right now because at the worst it wastes a bit of
space.  Even that is unlikely because (1) in most cases it makes sense
to display a heading, avoiding this issue, and (2) even if there is no
header typically a column will eventually be populated by callbacks.

But when we add support for hiding columns (including the flanking
separators) until they have a non-empty value, a callback may
introduce a non-empty value when there is no additional width
available to claim.  In this case, we need to avoid misalignment that
would occur from inserting the flanking separators for these updated
rows.

Re: https://github.com/pyout/pyout/pull/86#issuecomment-541697787