auto-width: Divide widths more evenly across columns
We don't use any inter-column context when automatically assigning
widths to columns.  If a column wants N characters and we have that
many free, we give it N characters.  Once there is no available width
left, we stop auto-adjusting widths.

If there are many columns or some columns have very long values but no
maximum width specified, this approach can quickly lead to columns not
being shown.  This situation will worsen when we add support for
hiding empty columns (as proposed in gh-86) because a callback may
give back a non-empty value well after the visible columns already
claimed all the width.

Rework the auto-width assignment to more evenly assign free space.
Ensure that each column gets at least one character, so that there is
at least some indication that a column is being truncated.  Then grow
each column together in step.  As noted in the code comment, the
current method could be optimized.  But punt on that until it is clear
that performance is an issue or at least until we known we're going to
stick with this new behavior.

It's worth highlighting that an important part of this change is
pulling out the logic for assigning widths to a separate method.  It's
very likely that the current method will need to be tweaked, and this
makes it easier to plug in different approaches, possibly even
exposing alternatives at the style level.

Closes #89.