elements: Allow values to be given as fraction of total width
This is particularly useful for setting the maximum of very long
fields.  You wouldn't want to set a column's maximum to a high integer
that risks exceeding the total width of the terminal.

Note the an important decision here is what the denominator is.  The
two obvious choices would be total table width or the total width
available to auto-width columns.  Let's go with the total table width,
which has the advantage of being easy for the caller to visualize and
of being easier to implement.  The main disadvantage is that the
spreading fractions across columns so that the values sum to one will
target more width than is actually available.  (Even when there are no
fixed columns, some width is claimed by the separators.)

Closes #85.