Extend test of StyleFunctionError's embedded traceback to Python 3
As of ea4156ee (BF: Raise StyleFunctionError with original traceback,
2018-02-11), when a transform function raises an exception, we embed
the original traceback so that we don't assume the caller handled the
exception in a way that preserves the original traceback.  The
motivating example for this was DataLad's exc_str() helper, which uses
extract_tb() underneath [*].

The regression test added in ea4156ee uses format_exception(), but
that only catches the issue under Python 2; under Python 3 the
original exception is included in the output.

Switch to using extract_tb() so that this regression test is effective
under Python 3 (which is important because we are about to drop Python
2 support).

[*] ... though, in light of Python 3's explicit chaining, exc_str() is
    arguably flawed.