interface: Use concurrent.futures.ThreadPoolExecutor
... instead of multiprocessing.dummy.Pool for asynchronous calls.

There wasn't too much thought that went into using
multiprocessing.dummy.Pool, but two factors were that (1) it worked on
Python 2 without any additional libraries and (2) it shared an
interface with multiprocessing.Pool, which left open the possibility
of exposing a way to let the user decide to use processes instead of
threads [*], which makes sense given that only they know the nature of
their task (e.g., whether it is IO-bound).

We've dropped support for Python 2, so no. 1 is no longer a concern,
and no. 2 is covered by ThreadPoolExecutor having a processor analog
ProcessPoolExecutor that shares the same Executor interface.

[*] ... though, to support processes, the callables would have to be
    picklable, so support would be limited to top-level functions
    supplied by the user.  Generator functions wouldn't work because
    we wrap those in another function.