interface: Relay exceptions from asynchronous workers
Now that we've switched to concurrent.futures we have access to
exceptions that occurred in the workers.  Within wait(), look at each
future as it completes, and record if it had an exception.  After all
the workers have finished, report all exceptions to stderr.

Note that logging these exceptions isn't a good option for reporting
them to the user because, to use logging usefully with pyout, the
output needs to be directed to a file, so the errors wouldn't be shown
on screen.  That leaves writing to stderr or to the provided stream as
the main options.  Let's use the provided stream because, as a
library, it seems best to not touch a stream unless the caller tells
us to.  The downside of writing to the provided stream is that both
Tabular's output and the error reporting happen on the same channel.

Closes #94.