interface: Show error if producer provides unexpected form
Now that we can show/raise exceptions for asynchronous callables, we
can provide better reporting when a producer is registered for
multiple columns but returns an atomic value.

As mentioned in the test, this unfortunately behaves differently
depending on whether the producer is a normal function or a generator.
For the normal function, the write happens in a callback, where
concurrent.futures will log any exceptions [*].  For a generator, we
need to wrap the generator to write each element, so the exception is
handled as any other exception in a producer.

One way to get consistent treatment would be to avoid registering a
callback in the case of a normal function, instead wrapping that as
well.  The problem with that is that, even if the user gives a
top-level, picklable function, we'll end up creating a producer that
can't be pickled.  This would complicate exposing ProcessPoolExecutor
as an alternative to ThreadPoolExecutor because there would be no
cases where we would feed in a producer that is picklable.

So, the inconsistency seems worth it to keep ProcessPoolExecutor an
option in some cases.

[*] ... unless the callable finishes _before_ add_done_callback()
    checks whether the future is done.  In that case, the exception
    won't be caught.