travis: Update Python 3.6 build to 3.8
I want to keep the matrix pretty slim until there's a particular
reason to expand it.  So if we're going to restrict the build to two
Python versions, we might as well test the bounds of what we support.

More concretely, an upcoming series needs to condition on Python 3.8,
and we'll want to cover that code path.