interface: Don't refer to async function for generator as a callback
In the non-generator case, the function that writes the updated values
is passed as the `callback` argument to apply_async().  For the
generator case, we need to process each value as it comes in, so we
wrap it in another function that calls this write function on each
yielded value.  This wrapped function is supplied as the main callable
to apply_async(), not the `callback` argument, so it's confusing to
refer to is as a callback.