interface: Support aborting on the first asynchronous exception
We now show exceptions, but they all come in at the end.  In some
cases, especially when debugging, it'd be better to abort on the first
failure.  Expose a `continue_on_failure` argument that can be set to
false to get that behavior.

The combination of relaying exceptions from asynchronous workers and
stopping on the first failure should reduce the need for gh-69, which
proposes serial execution for debugging purposes.  However, we should
still consider supporting execution in the main thread as an option
for callers that use sys.excepthook to drop into a debugger
(e.g., datalad's --dbg/--idbg switches).