Support an unbounded table width for non-interactive streams
If the stream isn't attached to a tty, it makes sense to let the
columns have as much space as needed, particularly in the case of the
output being redirected to a file.

When non-interactive, Stream.width already returns None.  Teach the
downstream code that None should be treated as unlimited available
width.

Closes #70.