setup.py: Make jsonschema a hard requirement
cc47ddfd (setup.py: Require at least version 3.0.0 of jsonschema,
2019-10-21) set a floor for the optional jsonschema dependency so that
the schema could use exclusiveMinimum and exclusiveMaximum.  However,
elements.validate() uses _any_ jsonschema it can import, so the
validation will fail (like this [0]) if pyout is installed with only
the core requirements and there is an older, incompatible jsonschema
exposed in the environment.

Resolve this issue by making jsonschema a hard requirement.  The other
option would be to teach validate() to check the particular version.
jsonschema isn't that heavy of a requirement, though, so it's probably
not worth the complication just so that we can keep it an optional
dependency.

[0]: https://github.com/pyout/pyout/pull/90#issuecomment-544163911

Reported-by: Yaroslav Halchenko <debian@onerussian.com>

Closes #100.