field: Provide more details if flank regexp does not match
The regexp in Flanks() is intended to always match.  gh-104 reported a
case in the wild where the assertion that there's a match failed.
Unfortunately, we don't know what the value that didn't match looked
like, and the failure now cannot be reproduced.  Expand the error
message so that, if this is triggered again, there's a better chance
of understanding what's going on from the exception message.