interface: Ignore unknown columns returned by callables
pyout depends on knowing the columns up front.  There are two ways
unknown keys can sneak in: as callable dict return values and as
callable column keys given to Tabular.__call__().  These unknown
columns can cause failures (e.g., gh-102) because the code assumes
that it is working with a fixed set of columns.

To deal with the first source, filter out unknown keys in dict results
returned by callables.  Another option would be to signal an error,
but it's not clear if there's much value in being strict here.  And
being more lenient allows callers to use some existing third-party
function that returns a dict that includes the information of interest
as well as other information.  If we signaled an error, the caller
would need to write a custom wrapper around the function to filter out
the keys.

Closes #102.