common: Don't let unknown columns in when starting callables
The previous commit dealt with one source of unknown columns: when a
callable returns a dictionary with extra keys.  The other way an
unknown column can be introduced is if Tabular.__call__() is passed a
value like

    {..., ("known", "unknown": lambda: ...)}

Filter out unknown columns when processing callables values.  Ignoring
these rather than signaling an error is in line with what is done if
(1) an unknown column is included by a callable in its result or (2)
Tabular.__call__() is given an unknown column with a non-callable
value.