Support expanding set of known columns
Lift the requirement that the set of columns be known up front.
Ideally this should be done in a single spot, because scattering
dict.get() and other handling throughout the code will be hard to
reason about and is likely to be a source of bugs.

Get close to that goal by making StyleFields.render() raise an
exception when it encounters an unknown column and then have _write(),
which importantly is working with an acquired lock, do the necessary
re-initializations.

Adjust some upstream spots to let unknown columns through, including
reverting the filtering of unknown columns from c21579bb (Merge pull
request #108 from kyleam/filter-unknown-cols, 2020-06-01).  Having the
upstream code itself register the new columns doesn't work well
because new columns can come in through plain values passed to
Tabular.__call__() and through the return values of asynchronous
functions.

Closes #49.