Fix intermittent failure in recently added test
The test added in 24829043 (interface: Try to coerce undocumented
types to a list, 2020-08-04) assumes that the column names inferred
from a dict row value will be in a particular order, which will fail
intermittently under Python 3.4 [*] and 3.5, where the dict
implementation does not preserve the order.  And even with later
Python versions, that's a bad assumption because the order-preserving
behavior of the new dict implementation is documented as an
implementation detail.

[*] Example: https://travis-ci.org/github/pyout/pyout/jobs/714928413