Detect changes in the terminal width
When taking the width from the terminal, the Writer class sets the
width once at setup time.  To detect changes in the width, instead
query TerminalStream.width (and thus blessings.Terminal.width) on each
write.

This is helpful in cases where the width of the terminal is being
expanded.  When decreasing it, the display is still disrupted by long
lines, if any, that need to be wrapped due to the shorter width.  But
the changes here shouldn't make that any worse.

Closes #71.