Content: Try rendering three times to work around unsolved bug
Content.__str__() renders all rows, catching a RedoContent error and
redoing the render.  This sequence is supposed to be sufficient to
account for adjustments in the field widths.  However, gh-114 reports
a scenario where, following the changes in f34696a7 (Detect changes in
the terminal width, 2020-08-17), the RedoContent exception makes it up
to the caller when the terminal is being resized.

With dandi-ls, I can trigger the same error, and it seems to be
related to having a header.  But I've been unsuccessful in isolating
the issue without dandi or in pinpointing why more than one redo is
needed in this case.  An additional redo appears sufficient, so do
that until a more proper fix is found.