field: Match newlines with Flanks regexp
The Flanks class chops off flanking whitespace before applying the
other post-format processors and then adds back the whitespace
afterward.  This prevents, for example, underlining the flanking
whitespace.

The regexp that matches the flanking whitespace should always match,
but it doesn't account for the fact that the caller could pass in a
value with newlines.  A field with newlines will break pyout in other
ways (e.g., line updates), so we may want to consider escaping values
to prevent this.  But either way, a value with newlines shouldn't
trigger a failure in the Flanks processor, so adjust the regexp to
allow for newlines.

Closes #104.
Closes #115.