tabular: Pass explicit stream to blessings.Terminal
The base Stream class sets the stream attribute to sys.stdout if the
stream argument isn't specified.  The TerminalStream subclass passes
the _argument_ as is to blessings.Terminal, letting it do its own
stream=None to stdout fallback.

This turns out to be problematic in the case of Jupyter notebooks.  In
this context sys.stdout is overridden to ipykernel.iostream.OutStream,
but blessings uses the original stdout, sys.__stdout__, as the stream.
As a result, there's no output when using pyout directly in Python
code of a notebook (or rather, it doesn't go to the expected stream).

Call blessings.Terminal with the stream attribute set by the base
class so that a sys.stdout override has an effect.  The isatty()
method of ipykernel.iostream.OutStream returns False, so there is no
styling and the "final" mode is used for rendering.  That's not
particularly nice, but it's better than no output at all.

Fixes #120