interface: Use "final" mode in Jupyter notebook process
If an external command that uses pyout (e.g., `dandi ls`) is called
from within a Jupyter notebook, the stream is still the usual stdout
stream, unlike the scenario from the previous commit.  However,
updates aren't effective, as described in gh-118.  On the other hand,
styling (bold, colors, ...) still works, and the width seems to be
reported at a fixed value of 80.

Of the available output modes, "final" is probably best in this
situation, though ideally there would be a specialized renderer for
Jupyter.  Try to detect if we're running from a notebook and, if so,
force non-interactive rendering in the "final" mode.  It doesn't look
like there is a reliable way to detect what frontend is hooked up to a
Jupyter kernel (by design) [1, 2].  Look for JPY_PARENT_PID in the
environment, and guess that we're probably in a notebook if that's
there.  On my end, this detects a Jupyter notebook while not
interfering with a standard IPython session in the terminal, but it
may need to be adjusted for different environments (e.g., spyder [1]).

This added check seems out of place in the Tabular.__init__().
Instead it should probably be somehow represented as another Stream
class.  However, there's more work needed to get nicer rendering, so
let's hold off on any large restructuring until then.  Another
questionable aspect is whether we should force non-interactive mode,
which loses styling and causes the use of unbound width rather than
the fixed value mentioned above, but it is consistent with the
rendering when pyout is used directly in the notebook (see the
previous commit).

Closes #118.

[1] https://stackoverflow.com/a/49368396
[2] https://stackoverflow.com/a/15412661